Added a new condition
Added the Input Condition. This condition requires the user to either
press or release a certain key to be met. Only keyboard inputs are
currently supported. Also added a new virtual function to
EnhancedTriggerBoxComponent.cs that can be overridden by components.
This function is named ResetComponent and is called after all responses
have been triggered and the trigger box is set to 'DoNothing'. It can be
used to reset variables or other objects. This is currently only being
used by InputCondition as a way to reset the triggered flag.