Numerous Updates
New option for AfterTrigger: Exit Responses. Allows the developer to
fire off responses when the object that entered the trigger box, leaves
the trigger box.
Updated all responses to take full advantage of the collision object
data.
Refactored a lot of the code in EnhancedTriggerBox.cs.
Included the new response as well as a demonstration of reusable trigger
boxes in the Examples scene.