Removed Demo Scene, refactored AudioResponse and fixed ModifyTransformResponse
BREAKING CHANGE: In AudioResponse there is a new enum that is used to split the response between modifying an Audio Source and playing a Sound Effect. By default, this enum is set to modify an Audio Source. If your ETB was previously playing a Sound Effect you'll need to go into the GUI and set the enum to playing a Sound Effect.

Demo Scene was out of date, was not working on below Unity5.4 and above Unity2017.1, so it has been removed. Examples scene has been improved to compensate for this.

AudioResponse has been refactored to make the GUI slightly easier to use by effectively splitting up the audio source modification fields and the special effect fields. Also introduced the ability to restart an audio clip.

Fixed ModifyTransform not correctly obtaining colliding object data. Also added an additive mode which allows you to add value to a transform's current values instead of just specifiying values.